pyproject.toml
[project]
name = "backend-ingres"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-google-genai>=2.1.12",
    "langchain-qdrant>=0.2.1",
    "langchain-text-splitters>=0.3.11",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
